+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.melvin.ongandroid">
+    package="com.melvin.ongandroid" >
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -11,16 +11,17 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:usesCleartextTraffic="true"
-        android:theme="@style/Theme.ONGAndroid">
+        android:theme="@style/Theme.ONGAndroid"
+        android:usesCleartextTraffic="true" >
         <activity
-            android:name=".view.NavigationActivity"
-            android:exported="false" />
+            android:name=".view.LoginActivity"
+            android:exported="true" />
         <activity
             android:name=".view.MainActivity"
-            android:exported="true">
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
