--- app/src/main/java/com/melvin/ongandroid/view/NavigationActivity.kt
@@ -1,46 +1,0 @@
-package com.melvin.ongandroid.view
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.Menu
-import androidx.activity.viewModels
-import androidx.fragment.app.viewModels
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.fragment.findNavController
-import androidx.navigation.ui.setupWithNavController
-import com.melvin.ongandroid.R
-import com.melvin.ongandroid.databinding.ActivityNavigationBinding
-import com.melvin.ongandroid.viewmodel.HomeViewModel
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class NavigationActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityNavigationBinding
-    private lateinit var navController: NavController
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityNavigationBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val navHostFragment =
-            supportFragmentManager.findFragmentById(R.id.navHostFragment) as NavHostFragment
-        navController = navHostFragment.findNavController()
-
-        //Bottom Navigation
-        binding.bottomNavigationView.setupWithNavController(navController)
-
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return navController.navigateUp() || super.onSupportNavigateUp()
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.bottom_menu, menu)
-        return true
-    }
-
-}
