+++ app/src/main/res/navigation/nav_login.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_login"
+    app:startDestination="@id/nav_log_in">
+
+    <fragment
+        android:id="@+id/nav_log_in"
+        android:name="com.melvin.ongandroid.view.login.FragmentLogIn"
+        android:label="fragment_log_in"
+        tools:layout="@layout/fragment_log_in" >
+        <action
+            android:id="@+id/action_fragmentLogIn_to_fragmentSignUp"
+            app:destination="@id/nav_sign_up" />
+    </fragment>
+    <fragment
+        android:id="@+id/nav_sign_up"
+        android:name="com.melvin.ongandroid.view.login.FragmentSignUp"
+        android:label="fragment_sign_up"
+        tools:layout="@layout/fragment_sign_up" />
+</navigation>
